requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
youtube-transcript-api==0.6.2
faiss-cpu
sentence-transformers==2.3.1
groq==0.4.2
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
numpy
